# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: upcontrol API
  version: 1.0.0
  description: |
    The front-facing contract for upcontrol. Generated into a strict Go server
    (oapi-codegen) and into TS types for `front/src/lib/api.d.ts`. Response
    shapes are pinned to `front/src/lib/mockData.ts` — a contract test
    (build-plan §2.7) checks each against golden JSON taken from the mocks.

    Rules from the plan (§4.4):
      - Version in the path (/v1), not a header.
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
        text reason. No `disabled`, no `navigate('/plan')`.
      - Pagination only on unbounded lists: incidents, logs. Monitors, channels
        and recipients are plan-bounded and returned whole.
      - `null` is never used for "no value": the field is omitted. "Zero is
        silence" — an absent field does not render.
      - All timestamps are RFC 3339 UTC.
      - `POST /i` and `POST /hooks/{provider}` are `x-hand-written: true`: they
        exist here for the /docs/api page but are implemented by hand (the
        generator's validation is forbidden on /i by §4.5).
servers:
  - url: http://localhost:5173
    description: local dev (front dev server proxies to ucapi)

# ---------------------------------------------------------------------------
# Common building blocks
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uc_session
    apiKey:
      type: apiKey
      in: header
      name: X-Upcontrol-Key
  responses:
...
    Unauthorized:
      description: No session, or the key is not ours.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PaymentRequired:
      description: |
        A plan gate blocked the action. The `upgrade` object carries a text
        reason; the client routes this to its upgrade prompt. No other
        paid-wall path exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist or belongs to another tenant.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "plan_limit_exceeded" }
            message: { type: string }
            upgrade:
              type: object
              description: Present only on 402. Drives the client's upgrade prompt.
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Free allows 3 HTTP checks."
                highlight:
                  type: string
                  description: Optional row to highlight on the pricing table.

    # --- account ------------------------------------------------------------
...
    HealthStatus:
      type: string
      enum: [ok, check, down, nodata]
    Monitor:
      type: object
      required: [id, type, name, target, status, interval]
      properties:
        id: { type: string, example: "mon_1" }
        type:
          type: string
          enum: [Website, Heartbeat]
        name: { type: string }
        target: { type: string, example: "https://example.com" }
        status: { $ref: "#/components/schemas/HealthStatus" }
        interval: { type: string, example: "5m" }
        keyword:
          type: string
          description: Optional assertion on the response body ("Page must contain").
        expiry:
          type: object
          description: >
            SSL and domain expiry, pre-baked for display. Absent for heartbeat,
            and absent entirely until one of the two has been read. Each half is
            present only when there is a date for it — "domain —" says we have
            not looked yet, not when the domain expires, and the row printed it
            under every check whose certificate we already had.
          properties:
            ssl: { type: string, example: "SSL Nov 2, in 84 days" }
            domain: { type: string, example: "domain Mar 14, in 217 days" }
    MonitorCreate:
      type: object
      required: [type, name, target, interval]
      properties:
        type:
          type: string
          enum: [Website, Heartbeat]
        name: { type: string }
        target: { type: string }
        keyword: { type: string }
        interval:
          type: string
          enum: ["1m", "5m", "15m", "30m", "1h"]
    MonitorPatch:
      type: object
      properties:
        name: { type: string }
        target: { type: string }
        keyword: { type: string }
        interval:
          type: string
          enum: ["1m", "5m", "15m", "30m", "1h"]
        paused: { type: boolean }

    # --- sources ------------------------------------------------------------
...
paths:
...
  /v1/monitors:
    get:
      tags: [monitors]
      summary: All monitors (plan-bounded, returned whole — no pagination).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Monitor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [monitors]
      summary: Create a monitor.
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/MonitorCreate" } },
          }
      responses:
        "201":
          description: Created
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/Monitor" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
...
  /v1/monitors/{id}:
    patch:
      tags: [monitors]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/MonitorPatch" } },
          }
      responses:
        "200":
          description: OK
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/Monitor" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [monitors]
      summary: |
        Delete a monitor. A check can be deleted (a typo in the third URL on Free
        must not occupy a slot forever); deletion asks first in the UI.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...